requirements.txt
asgiref==3.6.0
Django==4.2
django-extensions==3.2.1
graphql-core==3.2.3
psycopg2==2.9.5
python-dateutil==2.8.2
python-decouple==3.7
six==1.16.0
sqlparse==0.4.4
strawberry-django-plus==2.3.1
strawberry-graphql==0.174.0
strawberry-graphql-django==0.9.4
typing_extensions==4.5.0
tzdata==2023.3
